requirements_test.txt
pytest~=9.0
pytest-asyncio~=1.3
pytest-cov~=7.0
pytest-homeassistant-custom-component~=0.13
PyJWT~=2.10
PyTurboJPEG~=1.7
